split walker and hinted walker tests
