+++ packages/cspell-trie-lib/src/lib/walker/walker.test.ts
@@ -1,8 +1,6 @@
 import { walker } from './walker';
 import type { YieldResult } from './walkerTypes';
-import { hintedWalker } from './hintedWalker';
 import { orderTrie, createTriFromList } from '../trie-util';
-import { parseLinesToDictionary } from '../SimpleDictionaryParser';
 
 describe('Validate Util Functions', () => {
     test('Tests Walker', () => {
@@ -20,59 +18,6 @@ describe('Validate Util Functions', () => {
             goDeeper = text.length < 4;
         }
         expect(result).toEqual(sampleWords.filter((a) => a.length <= 4).sort());
-    });
-
-    test('Hinted Walker', () => {
-        const root = createTriFromList(sampleWords);
-        orderTrie(root);
-        // cspell:ignore joty
-        // prefer letters j, o, t, y before the others.
-        const i = hintedWalker(root, false, 'joty', undefined);
-        let goDeeper = true;
-        let ir: IteratorResult<YieldResult>;
-        const result: string[] = [];
-        while (!(ir = i.next({ goDeeper })).done) {
-            const { text, node } = ir.value;
-            if (node.f) {
-                result.push(text);
-            }
-            goDeeper = text.length < 4;
-        }
-        expect(result).toEqual('joy jowl talk lift walk'.split(' '));
-    });
-
-    test('Hinted Walker compounds', () => {
-        const dict = ['A*', '+a*', '*b*', '+c'];
-        const trie = parseLinesToDictionary(dict, { stripCaseAndAccents: true });
-        const i = hintedWalker(trie.root, false, 'a', undefined);
-        let goDeeper = true;
-        let ir: IteratorResult<YieldResult>;
-        const result: string[] = [];
-        while (!(ir = i.next({ goDeeper })).done) {
-            const { text, node } = ir.value;
-            if (node.f) {
-                result.push(text);
-            }
-            goDeeper = text.length < 2;
-        }
-        expect(result).toEqual(['A', 'Aa', 'Ab', 'Ac', 'b', 'ba', 'bb', 'bc']);
-    });
-
-    test('Hinted Walker compounds ignoreCase', () => {
-        const dict = ['A*', '+a*', '*b*', '+c'];
-        const trie = parseLinesToDictionary(dict, { stripCaseAndAccents: true });
-        const i = hintedWalker(trie.root, true, 'a', undefined);
-        let goDeeper = true;
-        let ir: IteratorResult<YieldResult>;
-        const result: string[] = [];
-        while (!(ir = i.next({ goDeeper })).done) {
-            const { text, node } = ir.value;
-            if (node.f) {
-                result.push(text);
-            }
-            goDeeper = text.length < 2;
-        }
-        expect(result).toEqual(['A', 'Aa', 'Ab', 'Ac', 'b', 'ba', 'bb', 'bc', 'a', 'aa', 'ab', 'ac']);
     });
 });
